backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
pandas==2.2.3
PyPDF2==3.0.1
openpyxl==3.1.2
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
SQLAlchemy==2.0.29
sqlalchemy-utils==0.42.1
pymysql==1.1.0
gunicorn==20.1.0
XlsxWriter==3.2.9
numpy<2
cryptography==42.0.5
pdfplumber==0.10.0
python-docx==0.8.12
